Due to API rate limits, Binance WebSocket was used for major coins, and Gate.io WebSocket was used for other minor coins that were missing.
The portfolio is now being updated in real time.

The transaction history menu is now more detailed. An additional transaction description is available for each transaction.

Coin sending process is now more consistent.

When purchasing coins from the market, live data is first retrieved from the Binance websocket. If the coin is not available on Binance, data is retrieved from the Coingecko website.